refactor: 비밀번호 변경은 reset -> update로 네이밍 수정 #35
