+++ src/route/path.js
@@ -33,7 +33,7 @@ export const PATH = {
     PROFILE: {
       ROOT: 'profile',
       UPDATE: 'update',
-      RESET_PW: 'reset-password',
+      UPDATE_PW: 'update-password',
     },
     MANAGE_MEMBERS: {
       ROOT: 'manage-members',
